requirements.txt
django
djangorestframework
Pillow
psycopg2-binary
python-decouple
python-slugify

